+++ tools/checkpoint_conversion/requirements.txt
@@ -1,0 +1,1 @@
+transformers
